Combine all the classes.
